# contracts/platform-v1/openapi.yaml
openapi: 3.1.0
info:
  title: Insight Platform API
  version: 1.0.0
  description: >-
    Current insight.platform/v1 public contract. Production L4-L6 qualification remains a
    separately reported deployment gate and does not change this repository contract status.
...
servers:
  - url: /v1
paths:
...
  /model-configuration:
    get:
      operationId: getModelConfiguration
      x-insight-operation-kind: query
      x-insight-command-receipt: false
      x-insight-permission: model.read
      x-insight-maximum-request-bytes: 16384
      responses:
        "200":
          description: Current authenticated model configuration result
          headers:
            Cache-Control: {$ref: "#/components/headers/PrivateNoStore"}
          content:
            application/json:
              schema: {$ref: "./schemas/model-configuration.schema.json#/$defs/ModelConfigurationCatalogViewV2"}
        "400": {$ref: "#/components/responses/PrivateApiProblem"}
        "401": {$ref: "#/components/responses/PrivateApiProblem"}
        "403": {$ref: "#/components/responses/PrivateApiProblem"}
        "404": {$ref: "#/components/responses/PrivateApiProblem"}
        "409": {$ref: "#/components/responses/PrivateApiProblem"}
        "503": {$ref: "#/components/responses/PrivateApiProblem"}
  /model-configuration/resources:
    get:
      operationId: listModelConfigurationResources
      x-insight-operation-kind: query
      x-insight-command-receipt: false
      x-insight-permission: model.read
      x-insight-maximum-request-bytes: 16384
      description: Current tenant Model resources in ascending ID order, with at most 25 items. The optional after ID is a keyset boundary, not a snapshot or authorization token. Current ModelRead is checked on every page.
      parameters:
        - {name: kind, in: query, required: true, schema: {enum: [model_provider, model_profile]}}
        - {name: after, in: query, schema: {$ref: "#/components/schemas/PlatformResourceId"}}
      responses:
        "200":
          description: Current authenticated model configuration result
          headers:
            Cache-Control: {$ref: "#/components/headers/PrivateNoStore"}
          content:
            application/json:
              schema: {$ref: "./schemas/model-configuration.schema.json#/$defs/ModelConfigurationResourcePageV1"}
        "400": {$ref: "#/components/responses/PrivateApiProblem"}
        "401": {$ref: "#/components/responses/PrivateApiProblem"}
        "403": {$ref: "#/components/responses/PrivateApiProblem"}
        "404": {$ref: "#/components/responses/PrivateApiProblem"}
        "409": {$ref: "#/components/responses/PrivateApiProblem"}
        "503": {$ref: "#/components/responses/PrivateApiProblem"}
...
  /model-default:
    get:
      operationId: getModelDefault
      summary: Read the tenant-owned exact default Model Deployment pointer
      tags: [Resources]
      x-insight-permission: model.read
      responses:
        "200":
          description: Current Tenant version and configured default, including an explicit null when unset.
          headers:
            ETag: {schema: {type: string, maxLength: 128}}
            Cache-Control: {$ref: "#/components/headers/PrivateNoStore"}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ModelDefaultViewV1"}
        "401": {$ref: "#/components/responses/PrivateApiProblem"}
        "403": {$ref: "#/components/responses/PrivateApiProblem"}
        "503": {$ref: "#/components/responses/PrivateApiProblem"}
    put:
      operationId: setModelDefault
      summary: Set or explicitly clear the tenant default under its current ETag
      tags: [Resources]
      x-insight-permission: model.write
      x-insight-idempotency: tenant_principal_scoped_receipt
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SetModelDefaultRequestV1"}
      responses:
        "200":
          description: The current Tenant projection after applying or replaying the authorized command.
          headers:
            ETag: {schema: {type: string, maxLength: 128}}
            Cache-Control: {$ref: "#/components/headers/PrivateNoStore"}
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ModelDefaultViewV1"}
        "400": {$ref: "#/components/responses/PrivateApiProblem"}
        "401": {$ref: "#/components/responses/PrivateApiProblem"}
        "403": {$ref: "#/components/responses/PrivateApiProblem"}
        "404": {$ref: "#/components/responses/PrivateApiProblem"}
        "409": {$ref: "#/components/responses/PrivateApiProblem"}
        "412": {$ref: "#/components/responses/PrivateApiProblem"}
        "428": {$ref: "#/components/responses/PrivateApiProblem"}
        "503": {$ref: "#/components/responses/PrivateApiProblem"}
...
  /{resource_noun}/{resource_id}/deployments/{deployment_id}:
    get:
      operationId: getManagedResourceDeployment
      summary: Read one immutable exact Deployment closure
      tags: [Resources]
      x-insight-permission: resource.read
      parameters:
        - $ref: "#/components/parameters/ResourceNoun"
        - $ref: "#/components/parameters/ResourceId"
        - $ref: "#/components/parameters/DeploymentId"
      responses:
        "200": {$ref: "#/components/responses/DeploymentResponse"}
        "401": {$ref: "#/components/responses/ApiProblem"}
        "403": {$ref: "#/components/responses/ApiProblem"}
        "404": {$ref: "#/components/responses/ApiProblem"}
        "503": {$ref: "#/components/responses/ApiProblem"}
...
components:
  parameters:
    ResourceNoun:
      name: resource_noun
      in: path
      required: true
      schema:
        type: string
        enum: [agents, skills, capabilities, capability-implementations, contexts, context-implementations, models, model-providers, mcp-servers, policies, sandbox-runtimes, sandbox-packages, sandboxes]
    ResourceId:
      name: resource_id
      in: path
      required: true
      schema: {$ref: "#/components/schemas/PublicManagementResourceId"}
...
    DeploymentId:
      name: deployment_id
      in: path
      required: true
      schema: {$ref: "#/components/schemas/PublicManagementDeploymentId"}
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 255, pattern: "^[ -~]+$"}
  responses:
    PrivateApiProblem:
      description: A bounded authenticated management problem without sensitive values.
      headers:
        trace-id: {$ref: "#/components/headers/TraceId"}
        Cache-Control: {$ref: "#/components/headers/PrivateNoStore"}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ApiProblem"}
    ApiProblem:
      description: A bounded, stable public problem response.
      headers:
        trace-id: {$ref: "#/components/headers/TraceId"}
        Cache-Control:
          $ref: "#/components/headers/NoStore"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiProblem"
...
    DeploymentResponse:
      description: One immutable typed Deployment closure.
      headers:
        trace-id: {$ref: "#/components/headers/TraceId"}
        ETag: {schema: {type: string, minLength: 1, maxLength: 128}}
        Cache-Control: {$ref: "#/components/headers/PrivateNoStore"}
      content:
        application/json:
          schema: {$ref: "#/components/schemas/DeploymentViewV1"}
...
  headers:
    TraceId:
      required: true
      schema:
        $ref: "#/components/schemas/TraceId"
    NoStore:
      schema:
        type: string
        const: no-store
    PrivateNoStore:
      schema:
        type: string
        const: no-store, private, max-age=0
...
  schemas:
    PublicManagementResourceKind:
      type: string
      enum: [agent, skill, capability_interface, capability_implementation, context_source_interface, context_source_implementation, model_provider, model_profile, mcp_server, policy, sandbox_runtime, sandbox_package, sandbox_profile]
    PublicManagementResourceId:
      type: string
      pattern: "^(agt|skl|cap|cim|ctx|xim|mpr|mdl|mcp|pol|srt|spk|sxp)_[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
    PublicManagementVersionId:
      type: string
      pattern: "^(aif|arev|srev|cirev|cimp|xirev|ximp|mprev|mdrev|mrev|prev|srrev|sprev|sxrev)_[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
    PublicManagementDeploymentId:
      type: string
      pattern: "^(adep|skdep|cdep|xdep|mpdep|mdep|mcdep|pdep|sxdep)_[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
...
    SetModelDefaultRequestV1:
      type: object
      additionalProperties: false
      required: [schema_version, default_model]
      properties:
        schema_version: {const: 1}
        default_model:
          oneOf:
            - {$ref: "#/components/schemas/ModelExactDeploymentRef"}
            - {type: "null"}
    ModelDefaultViewV1:
      type: object
      additionalProperties: false
      required: [schema_version, tenant_id, default_model, version, etag]
      properties:
        schema_version: {const: 1}
        tenant_id: {type: string, pattern: "^ten_[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"}
        default_model:
          oneOf:
            - {$ref: "#/components/schemas/ModelExactDeploymentRef"}
            - {type: "null"}
        version: {type: integer, minimum: 1, maximum: 9223372036854775807}
        etag: {type: string, minLength: 1, maxLength: 128}
...
    DeploymentViewV1:
      type: object
      additionalProperties: false
      required: [schema_version, deployment_id, resource_id, resource_kind, resource_version_id, environment, closure_digest, closure, created_at, etag]
      properties:
        schema_version: {const: 1}
        deployment_id: {$ref: "#/components/schemas/PublicManagementDeploymentId"}
        resource_id: {$ref: "#/components/schemas/PublicManagementResourceId"}
        resource_kind: {$ref: "#/components/schemas/PublicManagementResourceKind"}
        resource_version_id: {$ref: "#/components/schemas/PublicManagementVersionId"}
        environment: {type: string, minLength: 1, maxLength: 128}
        closure_digest: {$ref: "#/components/schemas/Digest"}
        closure: {$ref: "#/components/schemas/DeploymentClosure"}
        created_at: {$ref: "#/components/schemas/UtcTimestamp"}
        etag: {type: string, minLength: 1, maxLength: 128}
...
    DeploymentClosure:
      $ref: ./schemas/deployment-closure.schema.json
...
    ModelExactDeploymentRef:
      type: object
      additionalProperties: false
      required: [deployment_id, resource_kind, deployment_digest]
      properties:
        deployment_id:
          type: string
          pattern: "^mdep_[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
        resource_kind: {const: model_deployment}
        deployment_digest: {$ref: "#/components/schemas/Digest"}
...
    PlatformResourceId:
      $ref: ./schemas/resource-id.schema.json
    Digest:
      $ref: ./schemas/nominal/digest.schema.json
    UtcTimestamp:
      $ref: ./schemas/nominal/utc-timestamp.schema.json
...
    ApiProblem:
      $ref: ./schemas/nominal/api-problem.schema.json
    TraceId:
      $ref: ./schemas/nominal/trace-id.schema.json
...